WLA-514: refector Welsh data mappper
